+++ paws.common/tests/testthat/test_iniutil.R
@@ -1,4 +1,6 @@
 test_that("Reads in default profile", {
+  # clear down cache
+  for (key in names(ini_cache)) rm(list = key, envir = ini_cache)
   content <- read_ini("data_ini")
   profile <- "default"
   expect_equal(content[[profile]][["arg1"]], "value1")
@@ -6,6 +8,8 @@ test_that("Reads in default profile", {
 })
 
 test_that("Reads in alternative profile", {
+  # clear down cache
+  for (key in names(ini_cache)) rm(list = key, envir = ini_cache)
   content <- read_ini("data_ini")
   profile <- "foo"
   expect_equal(content[[profile]][["arg1"]], "foo_value1")
@@ -13,12 +17,16 @@ test_that("Reads in alternative profile", {
 })
 
 test_that("Ignores lines starting with # and ;", {
+  # clear down cache
+  for (key in names(ini_cache)) rm(list = key, envir = ini_cache)
   content <- read_ini("data_ini")
   expect_true(!("ignore1" %in% names(content)))
   expect_true(!("ignore2" %in% names(content)))
 })
 
 test_that("Reads in profile with space in name", {
+  # clear down cache
+  for (key in names(ini_cache)) rm(list = key, envir = ini_cache)
   content <- read_ini("data_ini")
   profile <- "profile bar"
   expect_equal(content[[profile]][["arg1"]], "bar_value1")
@@ -26,6 +34,8 @@ test_that("Reads in profile with space in name", {
 })
 
 test_that("Reads in values with equal signs", {
+  # clear down cache
+  for (key in names(ini_cache)) rm(list = key, envir = ini_cache)
   content <- read_ini("data_ini")
   profile <- "equalsign"
   expect_equal(content[[profile]][["arg1"]], "value1==")
@@ -33,6 +43,8 @@ test_that("Reads in values with equal signs", {
 })
 
 test_that("Reads in values with spaces around the equal sign", {
+  # clear down cache
+  for (key in names(ini_cache)) rm(list = key, envir = ini_cache)
   content <- read_ini("data_ini")
   profile <- "spaces"
   expect_equal(content[[profile]][["arg1"]], "value1")
@@ -40,6 +52,8 @@ test_that("Reads in values with spaces around the equal sign", {
 })
 
 test_that("Reads in values with nested content", {
+  # clear down cache
+  for (key in names(ini_cache)) rm(list = key, envir = ini_cache)
   content <- read_ini("data_ini")
   profile <- "nested"
   expect_equal(content[[profile]][["nested1"]], list(arg1 = "value1", arg2 = "value2"))
@@ -48,8 +62,18 @@ test_that("Reads in values with nested content", {
 })
 
 test_that("Reads empty profiles", {
+  # clear down cache
+  for (key in names(ini_cache)) rm(list = key, envir = ini_cache)
   content <- read_ini("data_ini")
   for (profile in c("empty1", "empty2")) {
     expect_null(content[[profile]])
   }
+})
+
+test_that("Check cache", {
+  # clear down cache
+  for (key in names(ini_cache)) rm(list = key, envir = ini_cache)
+  content1 <- read_ini("data_ini")
+  content2 <- read_ini("data_ini")
+  expect_equal(content1, content2)
 })