+++ paws.common/NEWS.md
@@ -1,5 +1,6 @@
 # paws.common 0.6.2.9000
 * minor performance improvement for `read_ini`
+* cache `read_ini` for improved performance
 
 # paws.common 0.6.2
 * fix how `read_ini` reads empty profiles from ini files